requirements.txt
streamlit
langchain>=0.3.0
langchain-openai>=0.1.0
langchain-community>=0.3.0
langgraph
pydantic
faiss-cpu
pypdf
duckduckgo-search>=6.0.0
pandas
python-dotenv
requests
tiktoken
matplotlib
python-dotenv
tavily-python
sarvamai

